# swagger-training/api/swagger/swagger.yaml
swagger: "2.0"
info:
  description: This is a API for nodejs-training project
  version: 1.0.0
  title: Node.js Training API
  contact:
    email: dmitry_folomkin@epam.com
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
host: localhost:9000
basePath: /api
consumes:
  - application/json
produces:
  - application/json
...
schemes:
- http
paths:
...
  /cities/{id}:
    put:
      tags:
      - cities
      summary: edit existing city or add new one
      description: edit existing city or add new one
      operationId: putCity
      consumes:
      - application/json
      produces:
      - application/json
      parameters:
      - name: id
        in: path
        description: id of city to be edited
        required: true
        type: string
        minimum: 0
        format: int32
      - in: body
        name: city
        description: city item with changed fields
        required: false
        schema:
          $ref: '#/definitions/City'
      responses:
        200:
          description: existing item changed
          schema:
            $ref: '#/definitions/City'
        201:
          description: new item added
        500:
          description: error appeared
    delete:
      tags:
      - cities
      summary: delete one city
      description: delete one city
      operationId: deleteCity
      produces:
      - application/json
      parameters:
      - name: id
        in: path
        description: id of city to be deleted
        required: true
        type: string
        minimum: 0
        format: int32
      responses:
        200:
          description: item deleted
          schema:
            $ref: '#/definitions/City'
...
definitions:
  City:
    type: object
    required:
    - capital
    - country
    - id
    - lat
    - long
    - name
    properties:
      id:
        type: string
        format: int32
        example: "1"
      name:
        type: string
        example: Ryazan
      country:
        type: string
        example: Russia
      capital:
        type: string
        example: Moscow
      lat:
        type: string
        format: float
        example: "70.5"
      long:
        type: string
        format: float
        example: "100.5"
    example:
      country: Russia
      capital: Moscow
      name: Ryazan
      id: "1"
      lat: "70.5"
      long: "100.5"
...